Split the general settings audit and lock handlers into a mixin.
